# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountDashboardResponse": {
        "additionalProperties": false,
        "properties": {
          "orders": {
            "items": {
              "$ref": "#/components/schemas/PortfolioOrderResponse"
            },
            "title": "Orders",
            "type": "array"
          },
          "overview": {
            "$ref": "#/components/schemas/AccountOverviewResponse"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionSnapshotResponse"
            },
            "title": "Positions",
            "type": "array"
          }
        },
        "required": [
          "overview",
          "positions",
          "orders"
        ],
        "title": "AccountDashboardResponse",
        "type": "object"
      },
      "AccountOverviewResponse": {
        "additionalProperties": false,
        "properties": {
          "available_cash": {
            "title": "Available Cash",
            "type": "number"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "current_net_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Net Value"
          },
          "daily_profit": {
            "title": "Daily Profit",
            "type": "number"
          },
          "daily_return_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Return Rate"
          },
          "frozen_cash": {
            "title": "Frozen Cash",
            "type": "number"
          },
          "initial_capital": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Capital"
          },
          "market_value": {
            "title": "Market Value",
            "type": "number"
          },
          "net_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Value"
          },
          "open_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Date"
          },
          "operation_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Days"
          },
          "performance_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performance Date"
          },
          "position_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Ratio"
          },
          "today_trade_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Today Trade Count"
          },
          "total_asset": {
            "title": "Total Asset",
            "type": "number"
          },
          "total_profit": {
            "title": "Total Profit",
            "type": "number"
          },
          "total_return_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Return Rate"
          }
        },
        "required": [
          "total_asset",
          "available_cash",
          "frozen_cash",
          "market_value",
          "total_profit",
          "daily_profit",
          "captured_at"
        ],
        "title": "AccountOverviewResponse",
        "type": "object"
      },
      "AccountRefreshResultResponse": {
        "additionalProperties": false,
        "properties": {
          "captured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "last_refresh_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Attempt At"
          },
          "last_refresh_succeeded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Succeeded At"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "AccountRefreshResultResponse",
        "type": "object"
      },
...
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PortfolioOrderResponse": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "filled_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Price"
          },
          "filled_quantity": {
            "title": "Filled Quantity",
            "type": "integer"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "order_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Price"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stock_name": {
            "title": "Stock Name",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "order_id",
          "symbol",
          "stock_name",
          "direction",
          "quantity",
          "filled_quantity",
          "status"
        ],
        "title": "PortfolioOrderResponse",
        "type": "object"
      },
      "PositionSnapshotResponse": {
        "additionalProperties": false,
        "properties": {
          "available_quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Quantity"
          },
          "avg_cost": {
            "title": "Avg Cost",
            "type": "number"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "current_price": {
            "title": "Current Price",
            "type": "number"
          },
          "day_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Profit"
          },
          "market_value": {
            "title": "Market Value",
            "type": "number"
          },
          "profit_ratio": {
            "title": "Profit Ratio",
            "type": "number"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "stock_name": {
            "title": "Stock Name",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "stock_name",
          "quantity",
          "avg_cost",
          "current_price",
          "market_value",
          "profit_ratio",
          "captured_at"
        ],
        "title": "PositionSnapshotResponse",
        "type": "object"
      },
...
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "csrf_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csrf Token"
          },
          "identity_initialized": {
            "title": "Identity Initialized",
            "type": "boolean"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "authenticated",
          "identity_initialized"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SetupIdentityRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "SetupIdentityRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AniuSessionCookie": {
        "in": "cookie",
        "name": "aniu_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "基于股票交易的可进化智能体系统",
    "title": "Aniu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/aniu/account/dashboard": {
      "get": {
        "operationId": "get_account_dashboard_api_aniu_account_dashboard_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request rate limited"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upstream integration failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Account Dashboard",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/aniu/account/refresh": {
      "post": {
        "operationId": "refresh_account_cache_api_aniu_account_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRefreshResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request rate limited"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upstream integration failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Refresh Account Cache",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/aniu/auth/login": {
      "post": {
        "operationId": "login_api_aniu_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid business request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/aniu/auth/logout": {
      "post": {
        "operationId": "logout_api_aniu_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid business request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/aniu/auth/session": {
      "get": {
        "operationId": "get_session_api_aniu_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid business request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/aniu/auth/setup": {
      "post": {
        "operationId": "setup_identity_api_aniu_auth_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid business request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Setup Identity",
        "tags": [
          "Auth"
        ]
      }
    },
...
  }
}